Second stage of moving code around
