--- game_system/src/implementation/mod.rs
@@ -1,10 +1,0 @@
-//! Module containing the effective game implementation code.
-//!
-//! The code within this module is supported by the state machine structure
-//! defined at medici::state_machine.
-
-pub mod effect;
-pub mod entity;
-pub mod prototype;
-pub mod runtime;
-pub mod setup;
