+++ game_system/src/setup.rs
